requirements.txt
fastapi==0.109.2
pillow==10.2.0
numpy==1.24.4
torch==2.2.0
pybase64==1.3.2
torchmetrics==1.3.1
pytorch-lightning
ultralytics
